Rename package from user to client
