+++ src/main/java/de/wevsvirushackathon/coronareport/diary/DiaryEntry.java
@@ -13,7 +13,7 @@ import javax.persistence.ManyToOne;
 
 import de.wevsvirushackathon.coronareport.contactperson.ContactPerson;
 import de.wevsvirushackathon.coronareport.symptomes.Symptom;
-import de.wevsvirushackathon.coronareport.user.Client;
+import de.wevsvirushackathon.coronareport.client.Client;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
