+++ src/main/java/de/wevsvirushackathon/coronareport/symptomes/SymptomDiaryController.java
@@ -24,8 +24,8 @@ import de.wevsvirushackathon.coronareport.diary.DiaryEntry;
 import de.wevsvirushackathon.coronareport.diary.DiaryEntryDtoIn;
 import de.wevsvirushackathon.coronareport.diary.DiaryEntryDtoOut;
 import de.wevsvirushackathon.coronareport.diary.DiaryEntryRepository;
-import de.wevsvirushackathon.coronareport.user.Client;
-import de.wevsvirushackathon.coronareport.user.ClientRepository;
+import de.wevsvirushackathon.coronareport.client.Client;
+import de.wevsvirushackathon.coronareport.client.ClientRepository;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
